+++ src/Ilias/ILIAS_OBJECT_TYPES.mjs
@@ -1,0 +1,7 @@
+export const ILIAS_OBJECT_TYPES = Object.freeze({
+    category: "Category",
+    course: "Course",
+    group: "Group",
+    "scorm-learning-module": "Scorm learning module",
+    test: "Test"
+});