+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-05-04-1-beta
+
+Changes:
+
+- value filter
+
 ## v2023-05-01-1-beta
 
 Changes: